@@ -1,33 +1,39 @@
 buildscript {
   ext {
-    company = "logicgate"
-    springBootVersion = "2.1.2.RELEASE"
+      checkstyleVersion = "10.25.0"
+      commonsLangVersion = "3.17.0"
+      company = "logicgate"
+      dependencyManagementVersion = "1.1.7"
+      gradleWrapperVersion = "8.14.1"
+      hamcrestVersion = "3.0"
+      h2Version = "2.3.232"
+      springBootVersion = "3.5.0"
+      slf4jVersion = "2.0.17"
   }
   repositories {
     mavenCentral()
   }
   dependencies {
     classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}"
+    classpath "io.spring.gradle:dependency-management-plugin:${dependencyManagementVersion}"
   }
 }
 
-apply plugin: "application"
-apply plugin: "checkstyle"
-apply plugin: "distribution"
-apply plugin: "idea"
-apply plugin: "java"
-apply plugin: "org.springframework.boot"
+plugins {
+  id "java"
+  id "application"
+  id "checkstyle"
+  id "distribution"
+  id "idea"
+  id("org.springframework.boot") version "${springBootVersion}"
+  id("io.spring.dependency-management") version "${dependencyManagementVersion}"
+}
 
 group = "com.logicgate"
 version = "1.0.0"
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-jar {
-  baseName = "farm"
-  manifest {
-    attributes 'Implementation-Title': 'LogicGate Farm', 'Implementation-Version': version
-  }
+java {
+    sourceCompatibility = JavaVersion.VERSION_21
+    targetCompatibility = JavaVersion.VERSION_21
 }
 
 repositories {
@@ -35,14 +41,15 @@ repositories {
 }
 
 dependencies {
-  compile "com.h2database:h2:1.4.197"
-  compile "com.puppycrawl.tools:checkstyle:7.8.2"
-  compile "org.apache.commons:commons-lang3:3.8.1"
-  compile "org.slf4j:slf4j-api:1.7.25"
-  compile "org.springframework.boot:spring-boot-starter-actuator:${springBootVersion}"
-  compile "org.springframework.boot:spring-boot-starter-data-jpa:${springBootVersion}"
-  compile "org.springframework.boot:spring-boot-starter-web:${springBootVersion}"
-  testCompile "org.springframework.boot:spring-boot-starter-test:${springBootVersion}"
+  implementation "com.h2database:h2:${h2Version}"
+  implementation "com.puppycrawl.tools:checkstyle:${checkstyleVersion}"
+  implementation "org.apache.commons:commons-lang3:${commonsLangVersion}"
+  implementation "org.slf4j:slf4j-api:${slf4jVersion}"
+  implementation "org.springframework.boot:spring-boot-starter-actuator"
+  implementation "org.springframework.boot:spring-boot-starter-data-jpa"
+  implementation "org.springframework.boot:spring-boot-starter-web"
+  testImplementation "org.springframework.boot:spring-boot-starter-test"
+  testImplementation "org.hamcrest:hamcrest:${hamcrestVersion}"
 }
 
 idea {
@@ -50,16 +57,21 @@ idea {
     inheritOutputDirs = true
   }
 }
-
-mainClassName = "com.logicgate.farm.Application"
+application {
+    mainClass = "com.logicgate.farm.Application"
+}
 
 checkstyle {
   configFile = file("${project.rootDir}/checkstyle.xml")
-  toolVersion = "7.8.2"
+  toolVersion = "${checkstyleVersion}"
   ignoreFailures = false
   showViolations = true
 }
 
 wrapper {
-  gradleVersion = "4.10"
+  gradleVersion = "${gradleWrapperVersion}"
+}
+
+test {
+    useJUnitPlatform()
 }
