@@ -1,15 +1,15 @@
 package com.logicgate.farm.domain;
 
-import java.io.Serializable;
-import java.util.Optional;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+import java.io.Serializable;
+import java.util.Objects;
 
 @Entity(name = "animal")
 public class Animal implements Serializable {
@@ -73,8 +73,7 @@ public class Animal implements Serializable {
 
   @Override
   public boolean equals(Object obj) {
-    return Optional.ofNullable(obj).isPresent()
-      && (obj == this || (obj.getClass() == getClass() && ((Animal) obj).getId().equals(getId())));
+    return this == obj || (obj instanceof Animal that && Objects.equals(id, that.id));
   }
 
 }
